+++ src/vs/workbench/contrib/void/browser/void.contribution.ts
@@ -1,14 +1,19 @@
+
+// register keybinds
+import './registerActions.js'
+
 // register Settings
 import './registerConfig.js'
 
-// register Sidebar chat
-import './registerSidebar.js'
+// register inline diffs
+import './registerInlineDiffs.js'
 
 // register Posthog metrics
 import './registerMetrics.js'
 
+// register Sidebar chat
+import './registerSidebar.js'
+
 // register Thread History
 import './registerThreads.js'
 
-// register inline diffs
-import './registerInlineDiffs.js'
